refactor: reorder popup vue components
