+++ src/popup/App.vue
@@ -6,17 +6,17 @@ import { useAttemptsStore } from '@/stores/attempts'
 import Navbar from '@/components/Navbar/Navbar.vue'
 import AppearTransition from '@common/components/Transitions/AppearTransition.vue'
 import Popups from '@/components/Popups/Popups.vue'
-import GroupNamePopup from '@/components/Popups/GroupNamePopup.vue'
+import GroupNamePopup from '@/components/Popups/Group/GroupNamePopup.vue'
 import ConfirmPopup from '@/components/Popups/ConfirmPopup.vue'
 import NewPasswordPopup from '@/components/Popups/NewPasswordPopup.vue'
 import PasswordPopup from '@/components/Popups/PasswordPopup.vue'
-import BindGroupPopup from '@/components/Popups/BindGroupPopup.vue'
-import ChooseEmojiPopup from '@/components/Popups/ChooseEmojiPopup.vue'
-import ChooseImageIconPopup from '@/components/Popups/ChooseImageIconPopup.vue'
+import BindGroupPopup from '@/components/Popups/Group/BindGroupPopup.vue'
+import ChooseEmojiPopup from '@/components/Popups/Group/ChooseEmojiPopup.vue'
+import ChooseImageIconPopup from '@/components/Popups/Group/ChooseImageIconPopup.vue'
 import LinkMenuPopup from '@/components/Popups/LinkMenuPopup.vue'
-import GroupMenuPopup from '@/components/Popups/GroupMenuPopup.vue'
-import EditGroupNamePopup from '@/components/Popups/EditGroupNamePopup.vue'
-import FolderNamePopup from '@/components/Popups/FolderNamePopup.vue'
+import GroupMenuPopup from '@/components/Popups/Group/GroupMenuPopup.vue'
+import EditGroupNamePopup from '@/components/Popups/Group/EditGroupNamePopup.vue'
+import FolderNamePopup from '@/components/Popups/Folder/FolderNamePopup.vue'
 
 onMounted(async () => {
     handleDarkTheme()
