+++ src/popup/components/Popups/Group/BindGroupPopup.vue
