+++ src/popup/components/Popups/Group/ChooseEmojiPopup.vue
