+++ src/popup/components/Popups/Group/ChooseImageIconPopup.vue
