+++ src/popup/components/Popups/Group/GroupMenuPopup.vue
