+++ src/popup/components/Views/FolderView/FolderView.vue
