+++ src/popup/router/routes.ts
@@ -3,7 +3,7 @@ import GroupView from '@/components/Views/GroupView/GroupView.vue'
 import SelectTabsView from '@/components/Views/SelectTabsView/SelectTabsView.vue'
 import GroupIconView from '@/components/Views/SetIconView/GroupIconView.vue'
 import GroupDetailsView from '@/components/Views/GroupDetailsView/GroupDetailsView.vue'
-import FolderView from '@/components/Views/GroupView/FolderView.vue'
+import FolderView from '@/components/Views/FolderView/FolderView.vue'
 
 export const routes = [
     { path: '/', name: 'main', component: MainView },
